+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 
 - widgets
     - button: add `decoupled` property (restrict visual feedback to values received from osc/midi messages)
-    - button: add client option `altTraversing` to allow one-way response to traversing gestures
+    - button: add client option `altTraversing` to allow one-way response to traversing gestures for toggle buttons
     - canvas: iOS: trigger `onTouch` when `event.force`, `event.altitudeAngle` or `event.azimuthAngle` updates
     - knob: add `snap-alt` mode (snap mode that allows jumping between range.min and range.max)
     - all: add `borderRadius` property